--- a/DragonSoul/Morchok.lua
+++ b/DragonSoul/Morchok.lua
@@ -9,6 +9,7 @@ mod:SetEncounterID(1292)
 mod:SetRespawnTime(30)
 
 local stompCount = 1
+local stompSkip = false
 
 --------------------------------------------------------------------------------
 -- Localization
@@ -74,9 +75,10 @@ end
 
 function mod:OnEngage()
 	self:Berserk(420) -- confirmed
-	self:Bar("stomp_boss", 11, L["stomp_boss"], L["stomp_boss_icon"])
+	self:CDBar("stomp_boss", 12.5, L["stomp_boss"], L["stomp_boss_icon"])
 	self:CDBar(103851, 57, L["blood"]) -- Black Blood, 57-64s
 	stompCount = 1
+	stompSkip = false
 end
 
 --------------------------------------------------------------------------------
@@ -84,7 +86,9 @@ end
 --
 
 function mod:SummonKohcrom(args)
-	self:CDBar("stomp_boss", 6, CL.other:format(self:SpellName(103414), self.displayName), L["stomp_boss_icon"]) -- Stomp, 6-12s
+	local stompCooldown = self:BarTimeLeft(L["stomp_boss"]) < 6 and 6 or 12
+	stompSkip = (stompCooldown == 6)
+	self:CDBar("stomp_boss", stompCooldown, CL.other:format(self:SpellName(103414), self.displayName), L["stomp_boss_icon"]) -- Stomp, 6-12s
 	self:MessageOld(args.spellId, "green")
 	self:StopBar(L["stomp_boss"])
 end
@@ -93,11 +97,11 @@ end
 function mod:BloodOver(_, _, _, spellId)
 	if spellId == 103851 then
 		self:CDBar(spellId, 74, L["blood"])
-		stompCount = 1
+		stompCount = 0
 		if self:Heroic() then
-			self:CDBar("stomp_boss", 15, CL.other:format(self:SpellName(103414), self.displayName), L["stomp_boss_icon"]) -- Stomp
+			self:CDBar("stomp_boss", 18, CL.other:format(self:SpellName(103414), self.displayName), L["stomp_boss_icon"]) -- Stomp
 		else
-			self:Bar("stomp_boss", 5, L["stomp_boss"], L["stomp_boss_icon"])
+			self:CDBar("stomp_boss", 18, L["stomp_boss"], L["stomp_boss_icon"])
 		end
 	end
 end
@@ -108,20 +112,25 @@ function mod:Stomp(args)
 			self:MessageOld("stomp_add", "red", nil, CL.other:format(args.spellName, self:SpellName(-4262)), args.spellId) -- "Stomp: Kohcrom"
 			self:StopBar(CL.other:format(args.spellName, self:SpellName(-4262))) -- "Stomp: Kohcrom"
 		else -- Since we trigger bars off morchok casts, we gotta make sure kohcrom isn't caster to avoid bad timers.
-			self:CDBar("stomp_add", self:Difficulty() == 5 and 6 or 5, CL.other:format(args.spellName, self:SpellName(-4262)), args.spellId) -- "Stomp: Kohcrom" 6sec after on 10 man hc, 5 sec on 25
+			if not stompSkip then
+				self:CDBar("stomp_add", 5, CL.other:format(args.spellName, self:SpellName(-4262)), args.spellId) -- "Stomp: Kohcrom"
+			end
+			stompSkip = false
 			self:MessageOld("stomp_boss", "red", nil, CL.other:format(args.spellName, self.displayName), args.spellId) -- "Stomp: Morchok"
+			stompCount = stompCount + 1
 			if stompCount < 4 then
 				self:CDBar("stomp_boss", 12, CL.other:format(args.spellName, self.displayName), args.spellId) -- "Stomp: Morchok"
 			else
 				self:StopBar(CL.other:format(args.spellName, self.displayName)) -- "Stomp: Morchok"
 			end
-			stompCount = stompCount + 1
 		end
 	else -- Not heroic, or Kohcrom isn't out yet, just do normal bar.
+		self:MessageOld("stomp_boss", "red", nil, args.spellId)
+		stompCount = stompCount + 1
 		if stompCount < 4 then
 			self:CDBar("stomp_boss", 12, args.spellId)
-			self:MessageOld("stomp_boss", "red", nil, args.spellId)
-			stompCount = stompCount + 1
+		else
+			self:StopBar(args.spellId)
 		end
 	end
 end