### Added respawn timer
### Added a stop for the black blood bar at the beginning of that phase
### Removed spawn timer bar for crystals
- This bar is superfluous & adds clutter which is a bit much on heroic mode with the duplicate bars due to the twin.
- The crystal spawn message & explosion timer bar (12.5 sec) remain.
### Improved stomp timings
- Timer for first stomp after the twin spawns depends on whether Morchok was able to perform a stomp _before_ the twin spawned.
- If Morchok's very first stomp for the encounter occurs _after_ the twin has spawned, the twin will <ins>not</ins> mimic it.
- After black blood has been used, Morchok uses stomp 4 times.
- **Before** black blood has been used, Morchok uses stomp 3 times (`stompCount = 1` on engage to skip one).
- I also removed the logic regarding Kochrom's mimicked stomp occurring after 6 sec on 10man & 5 sec on 25man. I've experienced myself & also found other logs in which both cases occur on 10man. It looks to be ~5 sec in both raid sizes but gets delayed if a crystal explodes just before the scheduled mimic stomp.