@funkydude 2025-03-06
Please re-add the old bar timings and use `:Retail()` to check for if they should be used. It's doubtful these have changed on Retail.