@funkydude 2025-03-10
> timings are the same on both Retail & Classic

Above an beyond, nice work!